requirements.txt
Django==5.1.7
djangorestframework==3.16.0
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
flake8  ==            7.2.0
django-filter   ==    25.1